requirements.txt
google-genai
feedparser
Pillow
requests